Shows reject card and voice correctly
If scene `entryConditions` are not met presents the reject card and voice to the user. Now also works when repeating the scene.

TODO: need to add support for alternate scene description text if the scene flags have been set. This will allow items to be picked up from the scene / levers flipped etc and the scene to show the correct description when it is revisited.