Added support for alternate scene description
Each scene can now specify an alternate voice and card to present to the user if alternate conditions are met. These conditions follow the same format as the reject condition. E.g. the room description can change to show that the user has picked up a "Red Key" and present the user with an alternate list of options that does not include the "Take Red Key" utterance.