Example game showcasing picking up and using a Red Key
The Red Key is located in the Hidden Room. Picking it up and travelling back to the starting room with the four doors will allow Door Number 2 to be opened with the key, allowing them access to a room filled with gold. If the player tries to open Door Number 2 early they will recieve a message saying the door is locked and that they need a Red Key.